(fix) better fix range handling (#160)
Fixes don't need to start at the same line where the error message is shown, and they don't need to be at the same line as the start of the error message. Therefore the previous handling of re-adding dedents was flawed. Instead, map the range to positions (line/character) and use them to find the correct dedent offsets.
Additionally add indentation after each newline a fix has because the fix doesn't know about the additional indentation.

To properly test fixes, the test setup was enhanced. If there's a Fixed.svelte, that one will be compared against a fixed version of Input.svelte.

Fixes #110
Fixes #156
Fixes #157